moved local screen/UI information into corresponding screen
